+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 -  Potentially breaking: Emit error when an attribute path is present in both `attributes` and `forward_attrs`. [#336](https://github.com/TedDriggs/darling/issues/336)
+-  Add `SpannedValue::into_inner` [#342](https://github.com/TedDriggs/darling/issues/342)
 
 ## v0.20.11 (March 28, 2025)